chore: Created constants file
